Created different User factories. Organized the files differently into seperate packages for cleaner final product.
