+++ src/main/java/entities/user/Student.java
@@ -1,4 +1,4 @@
-package entities;
+package entities.user;
 
 import java.util.Map;
